+++ oap-server/server-query-plugin/debugging-query-plugin/src/main/java/org/apache/skywalking/oap/query/debug/mqe/DebuggingMQERsp.java
@@ -31,5 +31,5 @@ public class DebuggingMQERsp {
     private final ExpressionResultType type;
     private final List<MQEValues> results;
     private final String error;
-    private final DebuggingTraceRsp execTrace;
+    private final DebuggingTraceRsp debuggingTrace;
 }